Enhances BackgroundTaskScheduler
