--- TweetNestKit/BackgroundTaskScheduler+WatchKit.swift
@@ -1,34 +1,0 @@
-//
-//  BackgroundTaskScheduler+WatchKit.swift
-//  BackgroundTaskScheduler+WatchKit
-//
-//  Created by Jaehong Kang on 2021/09/08.
-//
-
-#if canImport(WatchKit)
-import Foundation
-import WatchKit
-import UnifiedLogging
-
-extension BackgroundTaskScheduler {
-    @discardableResult
-    public nonisolated func handleBackgroundRefreshBackgroundTask(_ backgroundTasks: Set<WKRefreshBackgroundTask>) -> Set<WKRefreshBackgroundTask> {
-        guard let backgroundTask = backgroundTasks.first(where: { $0.userInfo as? NSString == Self.backgroundRefreshBackgroundTaskIdentifier as NSString }) else {
-            return []
-        }
-
-        Task.expiringDetached { expire in
-            let logger = Logger(subsystem: Bundle.tweetNestKit.bundleIdentifier!, category: "background-refresh")
-
-            backgroundTask.expirationHandler = {
-                logger.notice("Background task expired for: \(String(describing: backgroundTask.userInfo), privacy: .public)")
-                expire()
-            }
-
-            backgroundTask.setTaskCompletedWithSnapshot(await self.backgroundRefresh())
-        }
-
-        return [backgroundTask]
-    }
-}
-#endif
